+++ sdk/cosmos/azure-cosmos/src/test/java/com/azure/cosmos/CosmosBulkTest.java
@@ -5,6 +5,7 @@ package com.azure.cosmos;
 
 import com.azure.cosmos.implementation.ISessionToken;
 import com.azure.cosmos.implementation.guava25.base.Function;
+import com.azure.cosmos.implementation.guava25.collect.Lists;
 import com.azure.cosmos.models.CosmosBulkExecutionOptions;
 import com.azure.cosmos.models.CosmosBulkItemRequestOptions;
 import com.azure.cosmos.models.CosmosBulkOperations;
@@ -69,8 +70,8 @@ public class CosmosBulkTest  extends BatchTestBase {
 
         CosmosBulkExecutionOptions cosmosBulkExecutionOptions = new CosmosBulkExecutionOptions();
 
-        List<com.azure.cosmos.models.CosmosBulkOperationResponse<CosmosBulkAsyncTest>> bulkResponse = bulkContainer
-            .executeBulkOperations(cosmosItemOperations, cosmosBulkExecutionOptions);
+        List<com.azure.cosmos.models.CosmosBulkOperationResponse<CosmosBulkAsyncTest>> bulkResponse = Lists.newArrayList(bulkContainer
+            .executeBulkOperations(cosmosItemOperations, cosmosBulkExecutionOptions));
 
         assertThat(bulkResponse.size()).isEqualTo(totalRequest * 2);
 
@@ -102,8 +103,8 @@ public class CosmosBulkTest  extends BatchTestBase {
 
         CosmosBulkExecutionOptions cosmosBulkExecutionOptions = new CosmosBulkExecutionOptions();
 
-        List<com.azure.cosmos.models.CosmosBulkOperationResponse<Object>> bulkResponse = bulkContainer
-            .executeBulkOperations(cosmosItemOperations);
+        List<com.azure.cosmos.models.CosmosBulkOperationResponse<Object>> bulkResponse = Lists.newArrayList(bulkContainer
+            .executeBulkOperations(cosmosItemOperations));
 
         assertThat(bulkResponse.size()).isEqualTo(totalRequest);
 
@@ -150,8 +151,8 @@ public class CosmosBulkTest  extends BatchTestBase {
 
         CosmosBulkExecutionOptions cosmosBulkExecutionOptions = new CosmosBulkExecutionOptions();
 
-        List<com.azure.cosmos.models.CosmosBulkOperationResponse<TestDoc>> bulkResponse  = bulkContainer
-            .executeBulkOperations(deleteCosmosItemOperation, cosmosBulkExecutionOptions);
+        List<com.azure.cosmos.models.CosmosBulkOperationResponse<TestDoc>> bulkResponse  = Lists.newArrayList(bulkContainer
+            .executeBulkOperations(deleteCosmosItemOperation, cosmosBulkExecutionOptions));
 
         assertThat(bulkResponse.size()).isEqualTo(totalRequest);
 
@@ -193,8 +194,8 @@ public class CosmosBulkTest  extends BatchTestBase {
 
         CosmosBulkExecutionOptions cosmosBulkExecutionOptions = new CosmosBulkExecutionOptions();
 
-        List<com.azure.cosmos.models.CosmosBulkOperationResponse<Object>> bulkResponse  = bulkContainer
-            .executeBulkOperations(readCosmosItemOperations, cosmosBulkExecutionOptions);
+        List<com.azure.cosmos.models.CosmosBulkOperationResponse<Object>> bulkResponse  = Lists.newArrayList(bulkContainer
+            .executeBulkOperations(readCosmosItemOperations, cosmosBulkExecutionOptions));
 
         assertThat(bulkResponse.size()).isEqualTo(totalRequest);
 
@@ -239,8 +240,8 @@ public class CosmosBulkTest  extends BatchTestBase {
                 cosmosItemOperation.getPartitionKeyValue()));
         }
 
-        List<com.azure.cosmos.models.CosmosBulkOperationResponse<Object>> bulkResponse  = bulkContainer
-            .executeBulkOperations(replaceCosmosItemOperations);
+        List<com.azure.cosmos.models.CosmosBulkOperationResponse<Object>> bulkResponse  = Lists.newArrayList(bulkContainer
+            .executeBulkOperations(replaceCosmosItemOperations));
 
         assertThat(bulkResponse.size()).isEqualTo(totalRequest);
 
@@ -265,8 +266,8 @@ public class CosmosBulkTest  extends BatchTestBase {
     private void createItemsAndVerify(List<com.azure.cosmos.models.CosmosItemOperation> cosmosItemOperations) {
         CosmosBulkExecutionOptions cosmosBulkExecutionOptions = new CosmosBulkExecutionOptions();
 
-        List<com.azure.cosmos.models.CosmosBulkOperationResponse<Object>> bulkResponse = bulkContainer
-            .executeBulkOperations(cosmosItemOperations, cosmosBulkExecutionOptions);
+        List<com.azure.cosmos.models.CosmosBulkOperationResponse<Object>> bulkResponse = Lists.newArrayList(bulkContainer
+            .executeBulkOperations(cosmosItemOperations, cosmosBulkExecutionOptions));
 
         assertThat(bulkResponse.size()).isEqualTo(cosmosItemOperations.size());
 
@@ -328,8 +329,8 @@ public class CosmosBulkTest  extends BatchTestBase {
             cosmosItemOperations.add(CosmosBulkOperations.getReplaceItemOperation(
                 testDocToReplace.getId(), testDocToReplace, new PartitionKey(this.partitionKey1), firstReplaceOptions));
 
-            List<com.azure.cosmos.models.CosmosBulkOperationResponse<Object>> bulkResponses = bulkContainer
-                .executeBulkOperations(cosmosItemOperations);
+            List<com.azure.cosmos.models.CosmosBulkOperationResponse<Object>> bulkResponses = Lists.newArrayList(bulkContainer
+                .executeBulkOperations(cosmosItemOperations));
 
             assertThat(bulkResponses.size()).isEqualTo(cosmosItemOperations.size());
 
@@ -352,8 +353,8 @@ public class CosmosBulkTest  extends BatchTestBase {
             cosmosItemOperations.add(CosmosBulkOperations.getReplaceItemOperation(
                 testDocToReplace.getId(), testDocToReplace, new PartitionKey(this.partitionKey1), replaceOptions));
 
-            List<com.azure.cosmos.models.CosmosBulkOperationResponse<Object>> bulkResponses = bulkContainer
-                .executeBulkOperations(cosmosItemOperations);
+            List<com.azure.cosmos.models.CosmosBulkOperationResponse<Object>> bulkResponses = Lists.newArrayList(bulkContainer
+                .executeBulkOperations(cosmosItemOperations));
 
             assertThat(bulkResponses.size()).isEqualTo(cosmosItemOperations.size());
 
@@ -384,8 +385,8 @@ public class CosmosBulkTest  extends BatchTestBase {
                 readOptions));
             cosmosItemOperations.add(CosmosBulkOperations.getCreateItemOperation(testDocToCreate, new PartitionKey(this.partitionKey1)));
 
-            List<com.azure.cosmos.models.CosmosBulkOperationResponse<Object>> bulkResponses = bulkContainer
-                .executeBulkOperations(cosmosItemOperations);
+            List<com.azure.cosmos.models.CosmosBulkOperationResponse<Object>> bulkResponses = Lists.newArrayList(bulkContainer
+                .executeBulkOperations(cosmosItemOperations));
 
             assertThat(bulkResponses.size()).isEqualTo(cosmosItemOperations.size());
 
@@ -496,7 +497,7 @@ public class CosmosBulkTest  extends BatchTestBase {
 
         operations.add(CosmosBulkOperations.getCreateItemOperation(anotherTestDocToCreate, new PartitionKey(this.partitionKey1)));
 
-        List<com.azure.cosmos.models.CosmosBulkOperationResponse<Object>> bulkResponses = bulkContainer.executeBulkOperations(operations);
+        List<com.azure.cosmos.models.CosmosBulkOperationResponse<Object>> bulkResponses = Lists.newArrayList(bulkContainer.executeBulkOperations(operations));
 
         assertThat(bulkResponses.size()).isEqualTo(operations.size());
 
@@ -536,7 +537,7 @@ public class CosmosBulkTest  extends BatchTestBase {
         operations.add(
             CosmosBulkOperations.getDeleteItemOperation(this.TestDocPk1ExistingC.getId(), new PartitionKey(this.partitionKey1)));
 
-        List<com.azure.cosmos.models.CosmosBulkOperationResponse<Object>> bulkResponses = bulkContainer.executeBulkOperations(operations);
+        List<com.azure.cosmos.models.CosmosBulkOperationResponse<Object>> bulkResponses = Lists.newArrayList(bulkContainer.executeBulkOperations(operations));
 
         assertThat(bulkResponses.size()).isEqualTo(operations.size());
 
@@ -596,7 +597,7 @@ public class CosmosBulkTest  extends BatchTestBase {
 
         operations.add(CosmosBulkOperations.getReadItemOperation(this.TestDocPk1ExistingB.getId(), new PartitionKey(this.partitionKey1)));
 
-        List<com.azure.cosmos.models.CosmosBulkOperationResponse<Object>> bulkResponses = bulkContainer.executeBulkOperations(operations);
+        List<com.azure.cosmos.models.CosmosBulkOperationResponse<Object>> bulkResponses = Lists.newArrayList(bulkContainer.executeBulkOperations(operations));
         assertThat(bulkResponses.size()).isEqualTo(operations.size());
 
         assertThat(bulkResponses.get(0).getResponse().getStatusCode()).isEqualTo(HttpResponseStatus.CREATED.code());