+++ sdk/communication/azure-communication-callingserver/CHANGELOG.md
@@ -3,7 +3,6 @@
 ## 1.0.0-beta.5 (Unreleased)
 ### Features Added
 - Add support for TokenCredential authentication with CallingServerClientBuilder.
-- Upgraded `azure-communication-common` to 1.0.4
 
 ### Features Added
 
@@ -12,6 +11,8 @@
 ### Bugs Fixed
 
 ### Other Changes
+#### Dependency updates
+- Upgraded `azure-communication-common` to 1.0.4
 
 ## 1.0.0-beta.4 (2021-08-04)
 ### Bug Fixes